CODSOFT Task 3
Restaurant menu

A restaurant menu UI/UX project is the process of designing a menu that is both visually appealing and easy to use. The goal of a good restaurant menu is to make it easy for customers to find the dishes they want, and to encourage them to order more.

A good restaurant menu should be:

Visually appealing: The menu should be visually appealing and easy to read. The use of colors, fonts, and images should be carefully considered to create a visually appealing menu.

Easy to use: The menu should be easy to use and navigate. The customer should be able to easily find the necessary information and make their selections.

Categorized: The menu should be categorized, so that customers can easily find the dishes they are looking for. The categories should be clear and concise, and they should be consistent throughout the menu.